Changed namespace and refactor
