+++ src/api/Shrooms.Domain/Shrooms.Domain.csproj
@@ -144,8 +144,8 @@
     <Compile Include="ServiceExceptions\ServiceException.cs" />
     <Compile Include="Services\Administration\IAdministrationUsersService.cs" />
     <Compile Include="Services\Administration\AdministrationUsersService.cs" />
-    <Compile Include="Services\BlacklistStates\BlacklistService.cs" />
-    <Compile Include="Services\BlacklistStates\IBlacklistService.cs" />
+    <Compile Include="Services\BlacklistUsers\BlacklistService.cs" />
+    <Compile Include="Services\BlacklistUsers\IBlacklistService.cs" />
     <Compile Include="Services\Email\AdministrationUsers\AdministrationUsersNotificationService.cs" />
     <Compile Include="Services\Email\Kudos\IKudosNotificationService.cs" />
     <Compile Include="Services\Email\Kudos\KudosNotificationService.cs" />
